+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.1] - Strip HTML tags when printing reports
+
+### Added
+- strip HTML tags in Markdown report outputs, so that they paste properly (use --preserve-tags to not strip)
+
+
 ## [0.4.0] - Add search and replace, delete, targets commands
 
 ### Added